getNewVals now operational
I fixed a bug in getNewVals, where the close parenthesis for computing
the new uncertainties was in the wrong place, leading to error.

New test cases were added that helped debug things, and code was
added to handle exceptional cases in two of the test cases. The
problem arose from rounding error (e.g., should have been 1.0, but
instead was 0.999999999999999999998).