changed names and handles nans
I changed the name of binRvs.py to wgtdMean.py.
I also changed the name of the respective test routine.

One of the test routines would occasionally crash if there
were more NaNs than values (it computed the median). This worked
in local testing, but drawing from random samples resulted in it
failing on travis. This was fixed and it will now always work if
the intended output is as expected.